requirements.txt
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
fastapi==0.74.1
FastAPI-SQLAlchemy==0.2.1
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Mako==1.1.6
MarkupSafe==2.1.0
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.0
pyparsing==3.0.7
pytest==7.0.1
python-dotenv==0.19.2
requests==2.27.1
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
